[WIP] Add SelectionCollection test helper class
The helper class is a wrapper to a SelectionCollection pointer.
SelectionCollectionDataTest members now call the helper methods.

This will allow simpler testing of the cloning functionality
for SelectionCollections, by e.g. parsing inputs, cloning,
swapping the pointer the helper contains, and then running
tests identical to those that already exist.

refs #868

Change-Id: I312ffca935ef08ea98dfd230367fb5c89ea44934