RFC: Make ArrayRef a proper iterator
The iterator shouldn't be implicitly convertible to a pointer.

Most of the patch is independent of how to implement the
iterator and can be reviewed seperately.

I chose to implement the iterator using a strong typedef
to avoid all the boilerplate associated with creating an
iterator (requires a lot of operators and typedefs).

A strong type is a type used in place of another type to carry
specific meaning through its name. This avoids any error
where an implicit conversion would be an error, e.g.:
- assigning a length to a velocity
- passing two real arguments in wrong order

Question: Do we want a strong/named typedef library?
If so which one?
Possible choices (among others):
- joboccara/NamedType
- martinmoene/type
- foonathan/type_safe (in this patch)

I can write up a comparison of the choices if needed.

Fixes #2859

Change-Id: I80e2b6a75bca0973f550e0ab8fa62c622c330f90