[RFC] upper triangle constraint 3x3 matrices view
Implement constraint view on matrices where the upper triangle is
enforced to be zero. Accessing matrix elements in the upper triangle
will return a reference to a zero'd variable.

Instead of a full constraint matrix class similar to the Matrix3x3, I
have the impression that it is usually the view on the Matrix that we
would like to employ. For that the constraint matrix is much easier to
implement, as done in this patch.

Refs #2835

Change-Id: Ia8369d2d2a61fd2a79017d16759e0bd1499c2fa1