New tests for pdb2gmx and editconf
The editconf tests run only in single precision, as they do not
reproduce well in double precision. We don't have a good way to make
them reproduce without writing a much more complex checker. It's
better that we add test coverage for the normal precision than to have
no coverage until we have code that works for both kinds of precision
configuration.

Change-Id: If85768ec63469b148547506f77a84c1fa940ebf8