Remove gmx::Regex
We now always use std::regex. The wrapper type added the value of
standardizing regex flags, which was useful while we had to use only
the subset that was usable on all platforms. But the wrapper type made
it harder to use than needed, because e.g. they could not be stored in
a container. Removing gmx::Regex will create new opportunities.

Refs #2881

Change-Id: I97974527052b52c17bb26304b214e8b152dafb4b