Added Simulator comparison test
This change extracts the general machinery to compare two simulation
runs which are expected to yield identical results. Currently, one
such test is used to compare the results of a rerun to the results
of the original simulation run.

This also add a test to compare two runs of the same .tpr file, setting
an environment variable for one of the runs. This allows to test that
new features (implemented behind a feature flag using an environment
variable) which are expected to be equivalent to existing features
yield identical results. The test can be used for other feature flags,
but is currently used for the GMX_USE_MODULAR_SIMULATOR environment
variable, used by the upcoming modular simulator implementation. As
this implementation is not yet present, the current test is supposed
to always pass, as it simply runs the same .tpr file on the same code
path twice.

Change-Id: I5a36ce0c488ba4baf65ad4a137c6d1ea75fa6e60