[WIP] iForceSchedule + Separating NS-related Conditionality in do_force(...)
This is an attempt to enhance the readability of the do_force(...)
function by separating the NS conditionality. This aims to move
that conditionality one level up at the schedule decision state.

At runtime, the appropriate schedule is selected depending on
whether it is an NS step or not. Attempts have been made to
minimize code duplication.

Change-Id: Ie3c03a595553ab5c7f6ff5685b592c18ccabe425