Moving Force Calculation Functions
This is a part of a larger modification that aims to
introduce the concept of modular force schedules.

Functions relevant to this, mainly do_force_cutsVERLET(...)
and the functions called by it are moved into another dedicated
file.

In a series of commits, this functionality will be subsumed
into a dedicated force schedule, more implementations of which
would be introduced eventually.

Related: #2574

Change-Id: I7113b8cdef25d12d4d866ac1c1a2b7e198d49d95