Introduce IForceSchedule and ScheduleBuilder
This is the first in a series of patches and commits that aims to add
the force-schedule abstraction within GROMACS. This aims to enable the
development of custom schedules specific to method
(FMM/PME/RF/Multigrid), hardware architecture (accelerators) or
parallelization paradigms (MPI/task-based programming).

These implementations of schedules are expected to be easily
maintainable and modular. Each of the concrete implementations would
be able to optimally overlap communication, computation and reduction
tasks of the various processes involved.

Later, a selectSchedule(...) function would take in the execution context and
hopefully based on problem definition and hardware, would make an optimal
selection.

Change-Id: Ic05633af3a723436fb95bed530f787698f6a4bb7
Related: #2574
Future: #2594 #2595