Convert gmx dump to module
Used Options and more std::string.

Minor fixes and improvements to text. Removed the report of the known
bug, replacing it with a TODO.

Added eftRunInput to the filename option types understood by
FileNameOption machinery, so that tools that only operate on .tpr
files work properly. This is then used for gmx dump -s.

Refs #2877

Change-Id: I5bb1c73c39a4cdd820735f894d4cb55d171afd48