GPU halo exchange for position buffer
This patch has two functions: an initialization fn to create a
structure containing information required to perform the exchange
(including a map of indices involved in the exchange), and
the exchange itself. Init fn is activated with SCHEDULEPUREMULTIGPU
environment variable, pending implementation of proper scheduling
mechanism. Exchange fn is not yet called, awaiting availability of
GPU position buffer ops.

Implements part of #2816

Change-Id: I8e6473481ad4d943df78d7019681bfa821bd5798