Prepare for Jupyter notebook based acceptance tests.
The project road map includes an expression of acceptance criteria that
is incremental, testable, annotated, and can be used by a human
developer or automated test. This change supports a jupyter notebook
based expression of required functionality.

An additional Dockerfile is provided to facilitate environment
preparation for the Jupyter notebook server.

The jupyter notebook implementing and annotating feature tests appears
in a dependent change.

Change-Id: I6fdd04e9b8ffbb77fffefaa4aad7cd5b7aa81c43