Add grompp warning with use of GROMOS FF
Added OPLS-based simulation database entry to replace
the current GROMOS-based one, to avoid simulations triggering
warnings when this is not the intent of the test case.

This commit was cherry-picked from release-2019 and updated for master
branch. The latter makes more widespread use of the GROMOS test setup
that we wish to replace.

Split some PME tests so that we test normal PME with xyz PBC
separately from walls with xy PBC.

Removed unused duplicate files in testing.

Refs #2884

Change-Id: I55afcf7ea5b423fa27792acebe09ea520b475974