F buffer operations in CUDA (without PME reduction)
Splitting up patch 9170. This patch performs GPU buffer ops for force
buffer only, without any reduction of PME forces. Currently active for
all local force buffer ops. For non-local, we need
insertNonlocalGpuDependency mechanism in patch 9169. Non-local
buffer ops still done on CPU in this patch.

haveCpuForces is currently always true, which means that we
do an H2D transfer at the start of the buffer op. When the PME reduction is
integrated into the buffer ops it will be done directly on the GPU so
we can use (haveSpecialForces || haveCpuBondedWork) to only do this transfer when required.

I considered further splitting this to seperate the buffer ops and
reduction, but it doesn't really improve isolation because haveCpuForces
would still be required to determine if an extra reduction across GPU forces
and CPU forces is required.

TODO: set haveCpuForces as described above.

Implements part of #2817

Change-Id: Ice984425301d24bac1340e883698244489cd686e