NVIDIA Prototype (nondraft): GPU update&constraints plus D2D comms
This patch does not require review. It is a (slightly updated) copy of
draft patch 8858, but committed as non-draft patch to allow visibility
(pending availability of new sandbox branch).

This patch demonstrates NVIDIA's new developments. It is based
on the existing GPU buffer ops patch, and contains:
* GPU Update and Constraints
* Device MPI: PME/PP Gather and Scatter
  - Relatively straightforward solution using CUDA-Aware MPI
  * Device MPI: PP local/nonlocal exchanges
      - New functionality to pack device-buffers and exchange
            using CUDA-aware MPI
	        - Similar D2D exchanges also for Constraints Lincs part

All coordinate and force PCIe transfers have been removed on
regular timesteps - this is possible since all compute
and communication parts are now device-side.

Change-Id: Ifa242774449f6cadc4a437be0c7bd5d889b981cb