[RFC] Use con/destructor for t_topology
This changes the legacy topology datastructure to also use constructor
and destructor, adapts calling code to use the new construction.

As there is no possibility to have a deep copy operation, code that
previous performed shallow copies has been reworked to instead return
unique_ptr to the datatructure, moving the data instead of copying.

I'm open for suggestions on how to improve on this here.

This is needed to replace underlying datatypes that are used both in
t_topology and gmx_mtop_t

Refs #2833

Change-Id: I4034fa0f700052a5ad72afa2f6c1b25a346e28a2