Move MDModules initialization earlier
This will permit extending the functionality to also provide
command-line options and help descriptions for gmx tools.

Moved more functionality into Mdrunner::Builder

It also permits the Mdrunner constructor that is normally used to
prompt the construction of a properly formed MDModules. Removed a
comment about default initialization that is a normal fact of life
with constructors. Updating it would be wordy and no real value.

Refs #2877

Change-Id: I70126f297422a9076484083d461b0b003ae2e23b