Clean up mdrun help description
Details about how unix signals are implemented can go in the user
guide.

The nice level does not actually change with whether mpirun is used,
so we should not say so.

Mentions of "experimental" features that have been in the code for a
decade don't belong here.

Fixed a repeated word.

Preparation for #2877

Change-Id: Ib68ab15717e52a9155be869581c18ba21513f067