Getting rid of use_GPU(nbv) function
Using the useGPU() function build within nonbonded_verlet_t
struct

This is another patch in the cleaning efforts of sim_util.

The idea is to only have functions relevant to the force
schedules there so that it becomes easy to move it into
its own module with minimal merging pains.

Related: #2574

Change-Id: I0e0d1dc74aa18032dfb4ab84f3badbb83ef0365b