Remove isTemperatureCouplingStep(...)
This subroutine hides the checks on whether the temperature
coupling should be updated, making it harder to read the
sequence of events in the main loop. Also, the subroutine is
used only twice and both times some of the conditionals are
double checked (e.g. is Velocity-Verlet integrator is in use).

Change-Id: Idb4196c2f81095c09426b2798e79d2beea4f306e