Clarify the EnergyOutput I
The EnergyOutput::printStepToEnergyFile(...) method is responsible
for printing the current and average thermodynamic quantities as
well as anealing temperatures. The swich between what is printed
is done by passing enum argument to the function. Different
values of this argument lead to diferent, almost non-overlapping
code paths. Hence, it is natural to split the function. The
follow-up patch will make these into separate methods of the
EnergyOutput class, thus eliminating enum argument and making it
more clear what and how often is done.

Change-Id: Ifde9098cdddf85bc663f2943b2ee8778bc72a3ac