Create gmx::compat::string_view
This part of the C++-17 std will likely help in cases we want objects to
store views to std::strings instead of performing a copy of them.

Updated COPYING instructions and developer guide.

Excluded the external code from Doxygen checking.
Noted where to go for documentation in the file Doxygen.

Change-Id: Ibc75f32929d385159d27d0b7ba87f2133995e915