Improvements for energy output test.
EnergyOutput handles outputs to the text log file
as well as to the binary edr file with the number of
the printed energy terms determined by the simulation
setup. This commit adds the tests for the binry output
and increases the number of tested energy terms.

TODO Proper fix for memory leak in energy output.

Change-Id: I670b168a83c0a1883b821a4b317c0df3055183e8