Improve warning messages about duplicate definitions
Because the warning is fatal, the wording with "overriding" is
misleading. An override only occurs if the warning is suppressed.

Email on gmx-users made clear that the word "overriding" is not clear
enough for all users to understand that more than one definition was
provided. They do not necessarily know that grompp has already read in
the force field definition. Typically a self-contained molecule .itp
file might override definitions found in the force-field (whether
duplicate or different). In such cases, the user needs to consider
what their intentions are and to express them clearly.

Change-Id: Ie73f1dbe1233d6fd0df41628c770b8f2ad1f1b1b