Clarify pushing of interaction types
gcc was warning about possibly pushing atomNumber to a vector when its
value was not defined.  That was impossible because of the gmx_incons
earlier, but we can express the validity of the code much more clearly
to the reader/compiler.

Fixes CMAP implementation also

Extracted the common code to a function.

Change-Id: I74be61b125e05ff5505162440349fcef4ea50fdb