Updated clang -Weverything checks to clang-7
This is supposed to be used with the latest tested clang
compiler. That is now clang 7, but the logic is hard-coded to clang 6,
which is why some things only failed in post-submit testing where we
have such a compiler for the clang-cuda build.

Fixed issues only found with clang -Weverything.

Also fixed those found with clang-8, but as we do not yet use that in
Jenkins, they're commented out for now.

Change-Id: I83be1cb770d486cc81a453d4911361f06a056ce3