Use test with parameters for EnergyOutput
Some of the energy terms are printed out only if the specific
combination of thermostat/barostat/integrator is enabled. The
output also considerably different if it is from a rerun. This
commit changes the tests so that several possible combinations
of input parameters are tested.

This commit also fixes a possible bug when having more
acceleration groups than temperature coupling groups would
lead to under-allocation of temporary variable tmp_v.

Change-Id: Icc08e49a428b14ffd62d251662eb96184564b2c4