Introduce tri-state enums for restarts
Both the user choice for appending and the decision about how to
implement a restart are good to express as a three-way enumeration of
mutually-exclusive possibilities, rather than booleans.

Checkpoint restarts also need to consider whether KE quantities
need to be recomputed, which is now stored in t_ekinstate alongside
the data from which it was computed.

Together, these eliminate the ContinuationOptions struct.

Several booleans in implementation objects were renamed to be
consistent with the StartingBehavior enumeration values, so that the
code is easier to understand.

Moved the call to handleRestart out of updateFromCommandLine now that
it no longer needed to be there.

Used namespaces for handlerestart.cpp

Refs #2804, #2375

Change-Id: I1128b94e947c6ef355a1b137b8978faa227ab1a0