Rewrite starting behavior
Several necessary checks were deferred until the time the checkpoint
was read, which made this feature hard to implement and
understand. Having moved the code, if it will not be possible to
append, we can tell the user immediately.

Fixed a bug where mdrun -append would start from the .tpr
configuration when the checkpoint file was missing.

Opening the logfile in the non-appending cases is now
closely associated with the logic for how the restart
works.

Refs #2804, #2375

Change-Id: I83a846958619e72ddc9a5e9bae49a9b71221ad24