Bug fix: Constraints tests are not instantiated with CUDA
If the tests initialization macros is called with ValuesIn
function, the container provided should be filled. In the
current version, it is filled in the test SetUp function,
after the macros is processed. As a result, the empty
container is used in instantiation and no test are run.
Only CUDA-enabled builds are affected.

Change-Id: Ia2f38c9908d4b7d4277d282c01781b278eeeead0