Introduce Simulator base class
In view of having simulators differing based on user input rather
than a large single rigid simulator object that does everything,
this change introduces a ISimulator interface. The current Simulator
class (renamed to LegacySimulator), implements the interface.
Derived classes cannot be aggregate-initialized, so a constructor
has been added to LegacySimulator. This change paves the way to use
a builder for custom Simulator objects rather than the current
single implementation.

To avoid code repetition, this base class is not a pure interface,
but has the member variables of the current LegacySimulator and a
constructor. This avoids having to write a (largely duplicated)
constructor for each simulator implementation, and allows us to
use parameter packages within the builder to avoid having to write
out the 30+ variables outside of the ISimulator definition.

Change-Id: I6afa478e4d01f6d52b2ed7689acea51a0f69ad5c