Clean up renamed simulator.h file
This change is a follow-up to I9bc31588, in which the Integrator class
was renamed to Simulator, and integrator.{h,cpp} was accordingly
renamed to simulator.{h,cpp}. The former change did, however, miss
the include guard, a typedef, and two mentions of "integrator" in
comments. The current change addresses this.

Change-Id: I14d55eeb35c18f198d4123a7132fbee2366713df