Incorporate sample_restraint repository.
This commit transfers development of the sample MD plug-in
code from an external repository to the GROMACS repository.
It should not be considered to add or change functionality,
but serves as a base for future integrated development.

The sample_restraint subtree serves as the root of a
template repository that a researcher could start from
to develop gmxapi MD extension code. It is meant to be
self-contained and built against an existing GROMACS
installation. Moving it to a GROMACS repo subdirectory
makes it less accessible to extension developers, but
easier to test and maintain for GROMACS developers. Near
term follow-up should include the creation of an external
skeleton repository, downstream subtree repository, or
"cookie cutter" repository drawing core content from the
GROMACS subtree.

* Import project from GitHub
  - https://github.com/kassonlab/sample_restraint/releases/tag/v0.0.7.3
  - Bundled sources for external projects are omitted.
  - Bundled test data files are omitted.
* Updates for compatibility with the gmxapi Python package are deferred
  to future changes.

Refs #3027

Change-Id: I1033a4a74dbce1bf1f9220a31cc9c15055c5dbaf