Fix gmx dump
The program would fail because the filename for writing the output mdp
file was passed as valid empty string that could not be opened as an
output file.

Now only passes the string when it is not empty.

Added test and fixed possible floating point exception found by the
test.

Change-Id: I8b18fbc0dfc5642399935f32fb0b1999fad44c98